+++ project/src/const.ts
@@ -1,12 +1,3 @@
-export enum AppRoute {
-    Main = '/',
-    Login = '/login',
-    Film = '/films/:id',
-    MyList = '/my-list',
-    AddReview = '/films/:id/review',
-    Player = '/player/:id'
-}
-
 export enum AuthorizationStatus {
     Auth = 'AUTH',
     NoAuth = 'NO_AUTH',
