+++ project/src/routing/app-route.ts
@@ -1,0 +1,8 @@
+export enum AppRoute {
+    Main = '/',
+    Login = '/login',
+    Film = '/films/:id',
+    MyList = '/my-list',
+    AddReview = '/films/:id/review',
+    Player = '/player/:id'
+}
