+++ project/src/routing/redirect-service.ts
@@ -1,0 +1,14 @@
+export {
+  buildFilmPath,
+  buildFilmReviewPath,
+};
+
+type FilmId = string | number;
+
+function buildFilmPath(filmId: FilmId) {
+  return `/films/${filmId}`;
+}
+
+function buildFilmReviewPath(filmId: FilmId) {
+  return `${buildFilmPath(filmId)}/review`;
+}
